Extract skill directory content synchronization
Move atomic skill file writes, rollback capture, stale file cleanup, and excluded-file checks into SkillDirectoryContentSynchronizer. This isolates the file-content synchronization responsibility before splitting target detection and installation flow.
