+++ Packages/src/Editor/Infrastructure/SkillSetup/SkillDirectoryContentSynchronizer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f49120c9b8cc841288f87c533c8ccd68
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
